@@ -11,48 +11,55 @@
   },
   "dependencies": {
     "ajv": "^6.12.5",
-    "aws-sdk": "^2.344.0",
-    "bluebird": "^3.5.2",
+    "aws-sdk": "^2.761.0",
     "body-parser": "~1.18.3",
-    "cookie-parser": "~1.4.3",
-    "debug": "~4.1.0",
-    "express": "~4.16.4",
+    "convict": "^6.0.0",
+    "cookie-parser": "^1.4.5",
+    "debug": "^4.2.0",
+    "express": "^4.17.1",
     "express-jwt": "^6.0.0",
     "express-jwt-authz": "^1.0.0",
     "fs-extra": "^9.0.1",
-    "glob": "^7.1.2",
+    "glob": "^7.1.6",
     "is-invalid-path": "^0.1.0",
-    "jsonschema": "^1.1.1",
+    "jsonschema": "^1.2.6",
     "lodash": "^4.17.19",
-    "mongoose": "^5.7.5",
+    "mongoose": "^5.10.7",
     "morgan": "~1.9.1",
-    "multer": "^1.4.1",
+    "multer": "^1.4.2",
     "validator.ts": "^0.2.2"
   },
   "devDependencies": {
     "@types/bluebird": "^3.5.32",
-    "@types/body-parser": "^1.17.0",
-    "@types/cookie-parser": "^1.4.1",
-    "@types/express": "^4.16.0",
+    "@types/body-parser": "^1.19.0",
+    "@types/convict": "^5.2.1",
+    "@types/cookie-parser": "^1.4.2",
+    "@types/express": "^4.17.8",
     "@types/express-jwt": "0.0.40",
-    "@types/fs-extra": "^5.0.4",
-    "@types/glob": "^7.1.1",
+    "@types/fs-extra": "^5.1.0",
+    "@types/glob": "^7.1.3",
     "@types/jest": "^26.0.14",
-    "@types/joi": "^13.6.2",
-    "@types/lodash": "^4.14.117",
-    "@types/mongoose": "^5.2.20",
-    "@types/morgan": "^1.7.35",
-    "@types/multer": "^1.3.7",
+    "@types/joi": "^13.6.3",
+    "@types/jsonwebtoken": "^8.5.0",
+    "@types/lodash": "^4.14.161",
+    "@types/mongoose": "^5.7.36",
+    "@types/morgan": "^1.9.1",
+    "@types/multer": "^1.4.4",
+    "@types/node": "^14.11.2",
+    "@types/supertest": "^2.0.10",
+    "@typescript-eslint/eslint-plugin": "^4.6.0",
+    "@typescript-eslint/parser": "^4.6.0",
     "clear-require": "^2.0.0",
+    "eslint": "^7.12.1",
+    "eslint-config-prettier": "^6.15.0",
     "jest": "^26.4.2",
-    "jsonwebtoken": "^8.3.0",
+    "jsonwebtoken": "^8.5.1",
     "nyc": "^15.1.0",
-    "source-map-support": "^0.4.15",
-    "supertest": "^3.0.0",
+    "prettier": "^2.1.2",
+    "source-map-support": "^0.4.18",
+    "supertest": "^3.4.2",
     "ts-jest": "^26.4.0",
-    "tslint": "^5.11.0",
-    "tslint-config-airbnb": "^5.11.0",
-    "typescript": "^3.9.7"
+    "typescript": "^4.0.3"
   },
   "nyc": {
     "exclude": [
