@@ -1,60 +1,55 @@
-process.env.FILESYSTEM_STORAGE_ENABLED = true;
-process.env.FILESYSTEM_STORAGE_ROOT = '/tmp/filebank';
+/*import { Directory, File } from '../models';
+import { app } from '../app';
 
-const validation = require('../dist/validation/validation');
-const models = require('../dist/models').default;
-const app = require('../dist/app');
-const supertest = require('supertest');
-const jwt = require('jsonwebtoken');
-const fs = require('fs');
-const config = require('../dist/config').config;
-const chai = require('chai');
-const expect = chai.expect;
-const assert = chai.assert;
-chai.use(require('chai-as-promised'));
+import supertest from 'supertest';
+import jwt from 'jsonwebtoken';
+import fs from 'fs';
+
+import { config } from '../config';
 
 const allAccessToken = jwt.sign({ scope: 'filebank:read filebank:write filebank:delete' }, 'v3rys3cr3tK3y');
 
-const createDirectory = (obj) => new models.Directory(obj);
-const createFile = (obj) => new models.File(obj);
+export const createDirectory = (obj: any) => new Directory(obj).save();
+export const createFile = (obj: any) => new File(obj).save();
 
-const getRoutes = (fileVerifyCb) => {
+type BaseCallback = (...filePieces: string[]) => void;
+
+const getRoutes = (fileVerifyCb: BaseCallback) => {
   describe('GET route', () => {
     it('can list items in root directory', async () => {
       await Promise.all([
-        createDirectory({ name: 'test', refId: 'test' }).save(),
-        createDirectory({ name: 'test2', refId: 'test2' }).save(),
-        createDirectory({ name: 'test3', refId: 'test3' }).save(),
-        createFile({ name: 'test.txt', refId: 'test.txt', mimetype: 'application/octet-stream' }).save(),
-        createFile({ name: 'test2.txt', refId: 'test2.txt', mimetype: 'application/octet-stream' }).save(),
-        createFile({ name: 'test3.txt', refId: 'test3.txt', mimetype: 'application/octet-stream' }).save(),
+        createDirectory({ name: 'test', refId: 'test' }),
+        createDirectory({ name: 'test2', refId: 'test2' }),
+        createDirectory({ name: 'test3', refId: 'test3' }),
+        createFile({ name: 'test.txt', refId: 'test.txt', mimetype: 'application/octet-stream' }),
+        createFile({ name: 'test2.txt', refId: 'test2.txt', mimetype: 'application/octet-stream' }),
+        createFile({ name: 'test3.txt', refId: 'test3.txt', mimetype: 'application/octet-stream' }),
       ]);
       return supertest(app)
         .get('/')
         .set('Authorization', `Bearer ${allAccessToken}`)
         .expect('Content-Type', /json/)
         .expect(200)
-        .then(res => {
-          expect(res.body.length).to.equal(6);
+        .then((res) => {
+          expect(res.body.length).toEqual(6);
         });
     });
 
     it('can list items in subdirectory', async () => {
-      const directory = createDirectory({ name: 'subdir', refId: 'subdir' });
-      await directory.save();
+      const directory = await createDirectory({ name: 'subdir', refId: 'subdir' });
       await Promise.all([
-        createDirectory({ name: 'test', refId: 'test', parent: directory._id }).save(),
-        createDirectory({ name: 'test2', refId: 'test2', parent: directory._id }).save(),
-        createFile({ name: 'test.txt', refId: 'test.txt', mimetype: 'application/octet-stream', directory: directory._id }).save(),
-        createFile({ name: 'test2.txt', refId: 'test2.txt', mimetype: 'application/octet-stream', directory: directory._id }).save(),
+        createDirectory({ name: 'test', refId: 'test', parent: directory._id }),
+        createDirectory({ name: 'test2', refId: 'test2', parent: directory._id }),
+        createFile({ name: 'test.txt', refId: 'test.txt', mimetype: 'application/octet-stream', directory: directory._id }),
+        createFile({ name: 'test2.txt', refId: 'test2.txt', mimetype: 'application/octet-stream', directory: directory._id }),
       ]);
       return supertest(app)
         .get('/subdir')
         .set('Authorization', `Bearer ${allAccessToken}`)
         .expect('Content-Type', /json/)
         .expect(200)
-        .then(res => {
-          expect(res.body.length).to.equal(4);
+        .then((res) => {
+          expect(res.body.length).toEqual(4);
         });
     });
 
@@ -63,7 +58,7 @@ const getRoutes = (fileVerifyCb) => {
         .get('/nonexisting')
         .set('Authorization', `Bearer ${allAccessToken}`)
         .expect(404);
-      expect(await fileVerifyCb('nonexisting')).to.equal(false);
+      expect(await fileVerifyCb('nonexisting')).toEqual(false);
     });
 
     it('can get file content', async () => {
@@ -74,22 +69,22 @@ const getRoutes = (fileVerifyCb) => {
         .attach('file', `${__dirname}/resources/64-64.jpg`)
         .expect('Content-Type', /json/)
         .expect(200);
-      expect(await fileVerifyCb('64-64.jpg')).to.equal(true);
+      expect(await fileVerifyCb('64-64.jpg')).toEqual(true);
       await supertest(app)
         .get('/64-64.jpg')
         .set('Authorization', `Bearer ${allAccessToken}`)
         .expect(200)
         .expect('Content-Type', 'image/jpeg')
-        .then(res => {
-          assert(res.body instanceof Buffer, 'Return value must be buffer');
+        .then((res) => {
+          expect(res.body instanceof Buffer).toBeTruthy();
           const file = fs.readFileSync(`${__dirname}/resources/64-64.jpg`);
-          assert.equal(res.body.length, file.length, 'Content size must match.');
+          expect(res.body.length).toEqual(file.length);
         });
     });
   });
 };
 
-const postRoutes = (fileVerifyCb) => {
+const postRoutes = (fileVerifyCb: BaseCallback) => {
   describe('POST route', () => {
     it('can create directory on fs root', async () => {
       const dir = await supertest(app)
@@ -98,11 +93,11 @@ const postRoutes = (fileVerifyCb) => {
         .send({ name: 'test-1', type: 'directory' })
         .expect('Content-Type', /json/)
         .expect(200)
-        .then(res => {
-          assert(res.body.refId.startsWith('test-1'));
-          assert.equal(res.body.type, 'directory');
+        .then((res) => {
+          expect(res.body.refId.startsWith('test-1')).toBeTruthy();
+          expect(res.body.type).toEqual('directory');
         });
-      expect(await fileVerifyCb('test-1')).to.equal(true);
+      expect(await fileVerifyCb('test-1')).toEqual(true);
       return dir;
     });
 
@@ -113,9 +108,9 @@ const postRoutes = (fileVerifyCb) => {
         .send({ name: 'test-2', type: 'directory' })
         .expect('Content-Type', /json/)
         .expect(200)
-        .then(res => {
-          assert(res.body.refId.startsWith('test-2'));
-          assert.equal(res.body.type, 'directory');
+        .then((res) => {
+          expect(res.body.refId.startsWith('test-2')).toBe(true);
+          expect(res.body.type).toEqual('directory');
           return res.body;
         });
       const d = await supertest(app)
@@ -124,17 +119,17 @@ const postRoutes = (fileVerifyCb) => {
         .set('Authorization', `Bearer ${allAccessToken}`)
         .expect('Content-Type', /json/)
         .expect(200)
-        .then(res => {
-          assert(res.body.refId.startsWith('test-2/test-3'));
-          assert.equal(res.body.type, 'directory');
-          assert.equal(res.body.parent, subdir._id);
+        .then((res) => {
+          expect(res.body.refId.startsWith('test-2/test-3')).toBe(true);
+          expect(res.body.type).toEqual('directory');
+          expect(res.body.parent).toEqual(subdir._id);
         });
-      expect(await fileVerifyCb('test-2', 'test-3')).to.equal(true);
+      expect(await fileVerifyCb('test-2', 'test-3')).toEqual(true);
       return d;
     });
 
     it('cannot create directory with invalid metadata', async () => {
-      const dir = await supertest(app)
+      await supertest(app)
         .post('/')
         .set('Authorization', `Bearer ${allAccessToken}`)
         .send({
@@ -148,31 +143,31 @@ const postRoutes = (fileVerifyCb) => {
         })
         .expect('Content-Type', /json/)
         .expect(400);
-      expect(await fileVerifyCb('faildir')).to.equal(false);
-      return dir;
+      expect(await fileVerifyCb('faildir')).toEqual(false);
     });
 
     it('can create file on root dir', async () => {
+      const metadata = {
+        deepMeta: { thing: 1, another: 2, third: 'third' },
+        metaField: true,
+      };
       await supertest(app)
         .post('/')
         .set('Authorization', `Bearer ${allAccessToken}`)
-        .field('metadata', JSON.stringify({
-          deepMeta: { thing: 1, another: 2, third: 'third' },
-          metaField: true,
-        }))
+        .field('metadata', JSON.stringify(metadata))
         .field('type', 'file')
         .field('schema', 'Test')
         .attach('file', `${__dirname}/resources/fire.jpg`)
         .expect('Content-Type', /json/)
         .expect(200)
-        .then(res => {
-          assert.equal(res.body.mimetype, 'image/jpeg');
-          assert(res.body.refId.startsWith('fire.jpg'));
-          assert.equal(res.body.type, 'file');
-          assert.isObject(res.body.metadata, 'Metadata should be an object');
-          assert.equal(res.body.metadata.deepMeta.another, 2);
+        .then((res) => {
+          expect(res.body.mimetype).toEqual('image/jpeg');
+          expect(res.body.refId.startsWith('fire.jpg')).toBe(true);
+          expect(res.body.type).toEqual('file');
+          expect(res.body.metadata).toEqual(metadata);
+          expect(res.body.metadata.deepMeta.another).toEqual(2);
         });
-      expect(await fileVerifyCb('fire.jpg')).to.equal(true);
+      expect(await fileVerifyCb('fire.jpg')).toEqual(true);
     });
 
     it('can create file on sub dir', async () => {
@@ -182,31 +177,31 @@ const postRoutes = (fileVerifyCb) => {
         .send({ name: 'test-4', type: 'directory' })
         .expect('Content-Type', /json/)
         .expect(200)
-        .then(res => {
-          assert(res.body.refId.startsWith('test-4'));
-          assert.equal(res.body.type, 'directory');
-          return res.body;
+        .then((res) => {
+          expect(res.body.refId.startsWith('test-4')).toEqual(true);
+          expect(res.body.type).toEqual('directory');
         });
+      const metadata = {
+        deepMeta: { thing: 1, another: 2, third: 'third' },
+        metaField: true,
+      };
       await supertest(app)
         .post('/test-4')
         .set('Authorization', `Bearer ${allAccessToken}`)
-        .field('metadata', JSON.stringify({
-          deepMeta: { thing: 1, another: 2, third: 'third' },
-          metaField: true,
-        }))
+        .field('metadata', JSON.stringify(metadata))
         .field('type', 'file')
         .field('schema', 'Test')
         .attach('file', `${__dirname}/resources/fire.pdf`)
         .expect('Content-Type', /json/)
         .expect(200)
-        .then(res => {
-          assert.equal(res.body.mimetype, 'application/pdf');
-          assert(res.body.refId.startsWith('test-4/fire.pdf'));
-          assert.equal(res.body.type, 'file');
-          assert.isObject(res.body.metadata, 'Metadata should be an object');
-          assert.equal(res.body.metadata.deepMeta.third, 'third');
+        .then((res) => {
+          expect(res.body.mimetype).toEqual('application/pdf');
+          expect(res.body.refId.startsWith('test-4/fire.pdf')).toEqual(true);
+          expect(res.body.type).toEqual('file');
+          expect(res.body.metadata).toEqual(metadata);
+          expect(res.body.metadata.deepMeta.third).toEqual('third');
         });
-      expect(await fileVerifyCb('test-4', 'fire.pdf')).to.equal(true);
+      expect(await fileVerifyCb('test-4', 'fire.pdf')).toEqual(true);
     });
 
     it('can create file with different name', async () => {
@@ -223,14 +218,13 @@ const postRoutes = (fileVerifyCb) => {
         .attach('file', `${__dirname}/resources/fire.jpg`)
         .expect('Content-Type', /json/)
         .expect(200)
-        .then(res => {
-          assert.equal(res.body.mimetype, 'image/jpeg');
-          assert(res.body.refId.startsWith('another-fire.jpg'));
-          assert.equal(res.body.type, 'file');
-          assert.isObject(res.body.metadata, 'Metadata should be an object');
-          assert.equal(res.body.metadata.deepMeta.another, 2);
+        .then((res) => {
+          expect(res.body.mimetype).toEqual('image/jpeg');
+          expect(res.body.refId.startsWith('another-fire.jpg')).toEqual(true);
+          expect(res.body.type).toEqual('file');
+          expect(res.body.metadata.deepMeta.another).toEqual(2);
         });
-      expect(await fileVerifyCb('another-fire.jpg')).to.equal(true);
+      expect(await fileVerifyCb('another-fire.jpg')).toEqual(true);
     });
 
     it('cannot create file when schema is missing and required', async () => {
@@ -244,7 +238,7 @@ const postRoutes = (fileVerifyCb) => {
           .attach('file', `${__dirname}/resources/fire.pdf`)
           .expect('Content-Type', /json/)
           .expect(400);
-        expect(await fileVerifyCb('fire.pdf')).to.equal(false);
+        expect(await fileVerifyCb('fire.pdf')).toEqual(false);
         return promise;
       } finally {
         config.schemaRequired = false;
@@ -274,7 +268,7 @@ const postRoutes = (fileVerifyCb) => {
   });
 };
 
-const putMetaRoutes = (fileVerifyCb) => {
+const putMetaRoutes = (fileVerifyCb: BaseCallback) => {
   describe('PUT .meta route', () => {
     it('cannot create file with invalid metadata', async () => {
       const promise = await supertest(app)
@@ -289,7 +283,7 @@ const putMetaRoutes = (fileVerifyCb) => {
         .attach('file', `${__dirname}/resources/fire.pdf`)
         .expect('Content-Type', /json/)
         .expect(400);
-      expect(await fileVerifyCb('fire.pdf')).to.equal(false);
+      expect(await fileVerifyCb('fire.pdf')).toEqual(false);
       return promise;
     });
 
@@ -308,9 +302,9 @@ const putMetaRoutes = (fileVerifyCb) => {
         })
         .expect('Content-Type', /json/)
         .expect(200)
-        .then(res => {
-          assert(res.body.refId.startsWith('metadir'));
-          assert.equal(res.body.type, 'directory');
+        .then((res) => {
+          expect(res.body.refId.startsWith('metadir')).toEqual(true);
+          expect(res.body.type).toEqual('directory');
           return res.body;
         });
       return await supertest(app)
@@ -323,11 +317,11 @@ const putMetaRoutes = (fileVerifyCb) => {
         })
         .expect('Content-Type', /json/)
         .expect(200)
-        .then(res => {
-          assert.equal(res.body.metadata.metaField, false);
-          assert.equal(res.body.metadata.deepMeta.thing, 2);
-          assert.equal(res.body.metadata.deepMeta.another, 4);
-          assert.equal(res.body.metadata.deepMeta.third, 'fourth');
+        .then((res) => {
+          expect(res.body.metadata.metaField).toEqual(false);
+          expect(res.body.metadata.deepMeta.thing).toEqual(2);
+          expect(res.body.metadata.deepMeta.another).toEqual(4);
+          expect(res.body.metadata.deepMeta.third).toEqual('fourth');
         });
     });
 
@@ -346,12 +340,12 @@ const putMetaRoutes = (fileVerifyCb) => {
         })
         .expect('Content-Type', /json/)
         .expect(200)
-        .then(res => {
-          assert(res.body.refId.startsWith('metadir-2'));
-          assert.equal(res.body.type, 'directory');
+        .then((res) => {
+          expect(res.body.refId.startsWith('metadir-2')).toEqual(true);
+          expect(res.body.type).toEqual('directory');
           return res.body;
         });
-      return await supertest(app)
+      await supertest(app)
         .put(`/${dir._id}.meta`)
         .set('Authorization', `Bearer ${allAccessToken}`)
         .send({
@@ -377,8 +371,8 @@ const putMetaRoutes = (fileVerifyCb) => {
         .attach('file', `${__dirname}/resources/fire.pdf`)
         .expect('Content-Type', /json/)
         .expect(200)
-        .then((res) => res.body);
-      return await supertest(app)
+        .then(res => res.body);
+      await supertest(app)
         .put(`/${file._id}.meta`)
         .set('Authorization', `Bearer ${allAccessToken}`)
         .send({
@@ -389,10 +383,10 @@ const putMetaRoutes = (fileVerifyCb) => {
         .expect('Content-Type', /json/)
         .expect(200)
         .then((res) => {
-          assert.equal(res.body.metadata.metaField, false);
-          assert.equal(res.body.metadata.deepMeta.thing, 2);
-          assert.equal(res.body.metadata.deepMeta.another, 4);
-          assert.equal(res.body.metadata.deepMeta.third, 'fourth');
+          expect(res.body.metadata.metaField).toEqual(false);
+          expect(res.body.metadata.deepMeta.thing).toEqual(2);
+          expect(res.body.metadata.deepMeta.another).toEqual(4);
+          expect(res.body.metadata.deepMeta.third).toEqual('fourth');
         });
     });
 
@@ -410,8 +404,8 @@ const putMetaRoutes = (fileVerifyCb) => {
         .attach('file', `${__dirname}/resources/fire.pdf`)
         .expect('Content-Type', /json/)
         .expect(200)
-        .then((res) => res.body);
-      return await supertest(app)
+        .then(res => res.body);
+      await supertest(app)
         .put(`/${file._id}.meta`)
         .set('Authorization', `Bearer ${allAccessToken}`)
         .send({
@@ -424,8 +418,8 @@ const putMetaRoutes = (fileVerifyCb) => {
     });
 
     it('cannot update metadata non-existing item', async () => {
-      return await supertest(app)
-        .put(`/ffffffffffffffffffffffff.meta`)
+      await supertest(app)
+        .put('/ffffffffffffffffffffffff.meta')
         .set('Authorization', `Bearer ${allAccessToken}`)
         .send({
           deepMeta: { thing: 2, another: 4, third: 'fourth' },
@@ -437,7 +431,7 @@ const putMetaRoutes = (fileVerifyCb) => {
   });
 };
 
-const deleteRoutes = (fileVerifyCb) => {
+const deleteRoutes = (fileVerifyCb: BaseCallback) => {
   describe('DELETE route', () => {
     it('can delete directory on fs root', async () => {
       await supertest(app)
@@ -446,25 +440,23 @@ const deleteRoutes = (fileVerifyCb) => {
         .send({ name: 'deletethis', type: 'directory' })
         .expect('Content-Type', /json/)
         .expect(200)
-        .then(res => {
-          assert(res.body.refId.startsWith('deletethis'));
-          assert.equal(res.body.type, 'directory');
+        .then((res) => {
+          expect(res.body.refId.startsWith('deletethis')).toEqual(true);
+          expect(res.body.type).toEqual('directory');
         });
-      const del = await supertest(app)
+      await supertest(app)
         .delete('/deletethis')
         .set('Authorization', `Bearer ${allAccessToken}`)
         .expect(204);
-      expect(await fileVerifyCb('deletethis')).to.equal(false);
-      return del;
+      expect(await fileVerifyCb('deletethis')).toEqual(false);
     });
 
     it('cannot delete non-existing directory', async () => {
-      const del = await supertest(app)
+      await supertest(app)
         .delete('/nonexisting')
         .set('Authorization', `Bearer ${allAccessToken}`)
         .expect(404);
-      expect(await fileVerifyCb('nonexisting')).to.equal(false);
-      return del;
+      expect(await fileVerifyCb('nonexisting')).toEqual(false);
     });
 
     it('can delete file on fs root', async () => {
@@ -475,12 +467,12 @@ const deleteRoutes = (fileVerifyCb) => {
         .attach('file', `${__dirname}/resources/64-64.jpg`)
         .expect('Content-Type', /json/)
         .expect(200);
-      expect(await fileVerifyCb('64-64.jpg')).to.equal(true);
+      expect(await fileVerifyCb('64-64.jpg')).toEqual(true);
       await supertest(app)
         .delete('/64-64.jpg')
         .set('Authorization', `Bearer ${allAccessToken}`)
         .expect(204);
-      expect(await fileVerifyCb('64-64.jpg')).to.equal(false);
+      expect(await fileVerifyCb('64-64.jpg')).toEqual(false);
     });
 
     it('cannot delete non-existing file', async () => {
@@ -488,12 +480,12 @@ const deleteRoutes = (fileVerifyCb) => {
         .delete('/nonexisting.jpg')
         .set('Authorization', `Bearer ${allAccessToken}`)
         .expect(404);
-      expect(await fileVerifyCb('nonexisting.jpg')).to.equal(false);
+      expect(await fileVerifyCb('nonexisting.jpg')).toEqual(false);
     });
   });
 };
 
-const putRoutes = (fileVerifyCb) => {
+const putRoutes = (fileVerifyCb: BaseCallback) => {
   describe('PUT route', () => {
     it('can move file', async () => {
       await supertest(app)
@@ -504,16 +496,16 @@ const putRoutes = (fileVerifyCb) => {
         .attach('file', `${__dirname}/resources/64-64.jpg`)
         .expect('Content-Type', /json/)
         .expect(200);
-      expect(await fileVerifyCb('moveme.jpg')).to.equal(true);
+      expect(await fileVerifyCb('moveme.jpg')).toEqual(true);
       await supertest(app)
         .post('/')
         .set('Authorization', `Bearer ${allAccessToken}`)
         .send({ name: 'movehere', type: 'directory' })
         .expect('Content-Type', /json/)
         .expect(200)
-        .then(res => {
-          assert(res.body.refId.startsWith('movehere'));
-          assert.equal(res.body.type, 'directory');
+        .then((res) => {
+          expect(res.body.refId.startsWith('movehere')).toEqual(true);
+          expect(res.body.type).toEqual('directory');
         });
       await supertest(app)
         .put('/moveme.jpg')
@@ -521,11 +513,11 @@ const putRoutes = (fileVerifyCb) => {
         .send({ target: '/movehere' })
         .expect('Content-Type', /json/)
         .expect(200)
-        .then(res => {
-          assert(res.body.refId.startsWith('movehere/moveme.jpg'));
-          assert.equal(res.body.type, 'file');
+        .then((res) => {
+          expect(res.body.refId.startsWith('movehere/moveme.jpg')).toEqual(true);
+          expect(res.body.type).toEqual('file');
         });
-      expect(await fileVerifyCb('movehere', 'moveme.jpg')).to.equal(true);
+      expect(await fileVerifyCb('movehere', 'moveme.jpg')).toEqual(true);
     });
 
     it('cannot move file to nonexisting directory', async () => {
@@ -537,13 +529,13 @@ const putRoutes = (fileVerifyCb) => {
         .attach('file', `${__dirname}/resources/64-64.jpg`)
         .expect('Content-Type', /json/)
         .expect(200);
-      expect(await fileVerifyCb('moveme2.jpg')).to.equal(true);
+      expect(await fileVerifyCb('moveme2.jpg')).toEqual(true);
       await supertest(app)
         .put('/moveme2.jpg')
         .set('Authorization', `Bearer ${allAccessToken}`)
         .send({ target: '/movehere' })
         .expect(404);
-      expect(await fileVerifyCb('movehere', 'moveme2.jpg')).to.equal(false);
+      expect(await fileVerifyCb('movehere', 'moveme2.jpg')).toEqual(false);
     });
 
     it('can move directory', async () => {
@@ -553,9 +545,9 @@ const putRoutes = (fileVerifyCb) => {
         .send({ name: 'movethisdir', type: 'directory' })
         .expect('Content-Type', /json/)
         .expect(200)
-        .then(res => {
-          assert(res.body.refId.startsWith('movethisdir'));
-          assert.equal(res.body.type, 'directory');
+        .then((res) => {
+          expect(res.body.refId.startsWith('movethisdir')).toEqual(true);
+          expect(res.body.type).toEqual('directory');
         });
       await supertest(app)
         .post('/')
@@ -563,9 +555,9 @@ const putRoutes = (fileVerifyCb) => {
         .send({ name: 'moveheredir', type: 'directory' })
         .expect('Content-Type', /json/)
         .expect(200)
-        .then(res => {
-          assert(res.body.refId.startsWith('moveheredir'));
-          assert.equal(res.body.type, 'directory');
+        .then((res) => {
+          expect(res.body.refId.startsWith('moveheredir')).toEqual(true);
+          expect(res.body.type).toEqual('directory');
         });
       await supertest(app)
         .post('/movethisdir')
@@ -575,18 +567,18 @@ const putRoutes = (fileVerifyCb) => {
         .attach('file', `${__dirname}/resources/64-64.jpg`)
         .expect('Content-Type', /json/)
         .expect(200);
-      expect(await fileVerifyCb('movethisdir', 'moveme.jpg')).to.equal(true);
+      expect(await fileVerifyCb('movethisdir', 'moveme.jpg')).toEqual(true);
       await supertest(app)
         .put('/movethisdir')
         .set('Authorization', `Bearer ${allAccessToken}`)
         .send({ target: '/moveheredir' })
         .expect('Content-Type', /json/)
         .expect(200)
-        .then(res => {
-          assert(res.body.refId.startsWith('moveheredir/movethisdir'));
-          assert.equal(res.body.type, 'directory');
+        .then((res) => {
+          expect(res.body.refId.startsWith('moveheredir/movethisdir')).toEqual(true);
+          expect(res.body.type).toEqual('directory');
         });
-      expect(await fileVerifyCb('moveheredir', 'movethisdir')).to.equal(true);
+      expect(await fileVerifyCb('moveheredir', 'movethisdir')).toEqual(true);
     });
 
     it('can rename directory', async () => {
@@ -596,9 +588,9 @@ const putRoutes = (fileVerifyCb) => {
         .send({ name: 'movethisdir1', type: 'directory' })
         .expect('Content-Type', /json/)
         .expect(200)
-        .then(res => {
-          assert(res.body.refId.startsWith('movethisdir1'));
-          assert.equal(res.body.type, 'directory');
+        .then((res) => {
+          expect(res.body.refId.startsWith('movethisdir1')).toEqual(true);
+          expect(res.body.type).toEqual('directory');
         });
       await supertest(app)
         .post('/movethisdir1')
@@ -608,18 +600,18 @@ const putRoutes = (fileVerifyCb) => {
         .attach('file', `${__dirname}/resources/64-64.jpg`)
         .expect('Content-Type', /json/)
         .expect(200);
-      expect(await fileVerifyCb('movethisdir1', 'moveme.jpg')).to.equal(true);
+      expect(await fileVerifyCb('movethisdir1', 'moveme.jpg')).toEqual(true);
       await supertest(app)
         .put('/movethisdir1')
         .set('Authorization', `Bearer ${allAccessToken}`)
         .send({ target: '/moveheredir1' })
         .expect('Content-Type', /json/)
         .expect(200)
-        .then(res => {
-          assert(res.body.refId.startsWith('moveheredir1'));
-          assert.equal(res.body.type, 'directory');
+        .then((res) => {
+          expect(res.body.refId.startsWith('moveheredir1')).toEqual(true);
+          expect(res.body.type).toEqual('directory');
         });
-      expect(await fileVerifyCb('moveheredir1')).to.equal(true);
+      expect(await fileVerifyCb('moveheredir1')).toEqual(true);
     });
   });
 };
@@ -631,3 +623,4 @@ module.exports = {
   deleteRoutes,
   putRoutes,
 };
+*/
