@@ -1,8 +1,8 @@
-import mongoose = require('mongoose');
+import * as mongoose from 'mongoose';
 import { Directory, ItemType } from '../index';
 
 export interface DirectoryModel extends Directory, mongoose.Document {
-  parent: string | DirectoryModel;
+  parent: string | DirectoryModel;
 }
 
 const directorySchema = new mongoose.Schema({
