@@ -1,24 +1,25 @@
 import * as express from 'express';
-import models from '../models';
+import * as models from '../models/index';
 import * as _ from 'lodash';
 import { validateDirectory, validateDirectoryMeta, validateFile, validateFileMeta } from '../validation/validation';
-import { Item, ItemType, NotFoundError, UserInputError } from '../index';
-import * as multer from 'multer';
+import { Directory, File, ItemType, NotFoundError, UserInputError } from "../index";
+import multer from 'multer';
 import { DirectoryModel } from '../models/directory';
-import * as Bluebird from 'bluebird';
 import { FileModel } from '../models/file';
-import { config } from '../config';
-import DirectoryService = require('../service/DirectoryService');
-import PathService = require('../service/PathService');
-import expressJwtAuthz = require('express-jwt-authz');
-import jwtAuth = require('../util/jwt-auth');
+import config from '../config/config';
+import { buildTree, listItems } from '../service/DirectoryService';
+import { parsePath } from '../service/PathService';
+import jwtAuth from '../util/jwt-auth';
 
-const transportFactory = require('../transport/factory');
+import transportFactory from '../transport/factory';
+import { checkRole } from "../middleware/checkRole";
 const router: express.Router = express.Router();
 const storage = transportFactory.getInstance();
 const upload = multer();
 
-const nop = (req: express.Request, res:  express.Response, next: express.NextFunction) => {
+const authz = config.get("authz");
+
+const nop = (req: express.Request, res: express.Response, next: express.NextFunction) => {
   next();
 };
 
@@ -29,11 +30,11 @@ const handleFileUpload = async function (req: express.Request, res: express.Resp
   } catch (e) {
     metadata = req.body.metadata;
   }
-  const fileObject = {
+  const fileObject: File = {
     metadata,
     mimetype: req.body.mimetype || req.file ? req.file.mimetype : 'application/octet-stream',
-    name: req.body.name || req.file.originalname,
-    type: 'file',
+    name: req.body.name || req.file.originalname,
+    type: ItemType.File,
     refId: '',
   };
   const fileValidatorResult = validateFile(req.body.schema, fileObject);
@@ -47,14 +48,14 @@ const handleFileUpload = async function (req: express.Request, res: express.Resp
 
 router.get(
   ['/', '/*'],
-  config.authz.enabled ? jwtAuth : nop, config.authz.enabled ? expressJwtAuthz(config.authz.readScope) : nop,
+  [authz.enabled ? jwtAuth : nop, authz.enabled ? checkRole(_.split(authz.readScope, ',')) : nop],
   async (req: express.Request, res: express.Response) => {
     try {
-      const pathObj = await PathService.parsePath(req.params[0] || '/', true).populate();
+      const pathObj = await parsePath(req.params[0] || '/', true).populate();
       switch (pathObj.current ? pathObj.current.type : ItemType.Directory) {
         case ItemType.Directory:
           if (!pathObj.exists()) throw new NotFoundError();
-          res.json(await DirectoryService.listItems(pathObj.current as DirectoryModel));
+          res.json(await listItems(pathObj.current as DirectoryModel));
           break;
 
         case ItemType.File:
@@ -70,18 +71,18 @@ router.get(
   });
 
 router.put(
-  '/:id\.meta',
-  config.authz.enabled ? jwtAuth : nop, config.authz.enabled ? expressJwtAuthz(config.authz.writeScope) : nop,
+  "/:id.meta",
+  [authz.enabled ? jwtAuth : nop, authz.enabled ? checkRole(_.split(authz.writeScope, ',')) : nop],
   async (req: express.Request, res: express.Response) => {
     try {
-      const [dir, file] = await Bluebird.all([
+      const [dir, file] = await Promise.all([
         models.Directory.findById(req.params.id).exec(),
         models.File.findById(req.params.id).exec(),
       ]);
       if (!dir && !file) {
         throw new NotFoundError('Not found.');
       }
-      if (config.schemaRequired && !req.body.schema) {
+      if (config.get("schemaRequired") && !req.body.schema) {
         throw new UserInputError('Schema parameter is required.');
       }
       if (dir) {
@@ -113,16 +114,16 @@ router.put(
 
 router.post(
   ['/', '/*'],
-  config.authz.enabled ? jwtAuth : nop, config.authz.enabled ? expressJwtAuthz(config.authz.writeScope) : nop,
+  [authz.enabled ? jwtAuth : nop, authz.enabled ? checkRole(_.split(authz.writeScope, ',')) : nop],
   upload.single('file'),
   async (req: express.Request, res: express.Response) => {
     const path: string[] = _.filter((req.params[0] || '').split('/'), part => !_.isEmpty(part));
 
     try {
-      const dirs = await DirectoryService.buildTree(path);
+      const dirs = await buildTree(path);
       const currentDir = dirs.pop();
 
-      if (config.schemaRequired && !req.body.schema) {
+      if (config.get("schemaRequired") && !req.body.schema) {
         throw new UserInputError('Schema parameter is required.');
       }
 
@@ -165,7 +166,7 @@ router.post(
 
 router.put(
   ['/', '/*'],
-  config.authz.enabled ? jwtAuth : nop, config.authz.enabled ? expressJwtAuthz(config.authz.writeScope) : nop,
+  [authz.enabled ? jwtAuth : nop, authz.enabled ? checkRole(_.split(authz.writeScope, ',')) : nop],
   async (req: express.Request, res: express.Response) => {
     if (!req.body.target) {
       res.status(400).send('Missing target parameter.');
@@ -173,8 +174,8 @@ router.put(
     }
 
     try {
-      const sourcePath = await PathService.parsePath(req.params[0]).populate();
-      const targetPath = await PathService.parsePath(req.body.target).populate();
+      const sourcePath = await parsePath(req.params[0]).populate();
+      const targetPath = await parsePath(req.body.target).populate();
       if (!targetPath.basename) {
         throw new UserInputError('Target name empty.');
       }
@@ -185,12 +186,14 @@ router.put(
       switch (sourcePath.current.type) {
         case ItemType.Directory:
           const dir = sourcePath.current as DirectoryModel;
-          const targetRef = targetPath.current || _.last(targetPath.subTree) || { name: targetPath.basename };
-          const newDir = await storage.moveDirectory(sourcePath.current, targetRef, (oldItem: Item, newItem: Item) => {
+          const targetRef = targetPath.current || _.last(targetPath.subTree) || { name: targetPath.basename, parent: null };
+          const { directory: newDir, items } = await storage.moveDirectory(sourcePath.current, targetRef as Directory);
+          await Promise.all(items.map(async (change) => {
+            const { oldItem, newItem } = change;
             return newItem.type === ItemType.Directory ?
               models.Directory.findOneAndUpdate({ refId: oldItem.refId }, { $set: { refId: newItem.refId } }).exec() :
               models.File.findOneAndUpdate({ refId: oldItem.refId }, { $set: { refId: newItem.refId } }).exec();
-          });
+          }))
           dir.refId = newDir.refId;
           dir.name = newDir.name;
           await dir.save();
@@ -219,12 +222,12 @@ router.put(
 
 router.delete(
   ['/', '/*'],
-  config.authz.enabled ? jwtAuth : nop, config.authz.enabled ? expressJwtAuthz(config.authz.deleteScope) : nop,
+  [authz.enabled ? jwtAuth : nop, authz.enabled ? checkRole(_.split(authz.deleteScope, ',')) : nop],
   async (req: express.Request, res: express.Response) => {
-    const path: string[] = _.filter((req.params[0] || '').split('/'), part => !_.isEmpty(part));
+    // const path: string[] = _.filter((req.params[0] || '').split('/'), part => !_.isEmpty(part));
 
     try {
-      const deletePath = await PathService.parsePath(req.params[0]).populate();
+      const deletePath = await parsePath(req.params[0]).populate();
       if (!deletePath.current) {
         throw 'Not found.';
       }
@@ -250,4 +253,4 @@ router.delete(
     }
   });
 
-export = router;
+export default router;
