@@ -1,23 +1,34 @@
-import { Directory, File, FileStorage, Item, ItemType } from '../index';
+import { Directory, File, FileStorage, Item, ItemType, MoveDirectoryResponse } from "../index";
 import { DeleteObjectsRequest } from 'aws-sdk/clients/s3';
 import * as _ from 'lodash';
 import { S3StorageOptions } from '../config';
 import { EventEmitter } from 'events';
-import awsSdk = require('aws-sdk');
-import s3 = require('aws-sdk/clients/s3');
+import * as AWS from 'aws-sdk';
+import S3 from 'aws-sdk/clients/s3';
 import path = require('path');
+import config from '../config/config';
 
-awsSdk.config.setPromisesDependency(Promise);
-awsSdk.config.update({ region: 'eu-west-1' });
+const { s3, enabled } = config.get("storage");
+
+AWS.config.setPromisesDependency(Promise);
+AWS.config.update({ region: s3.clientOptions.region, s3ForcePathStyle: true });
 
 export class S3Storage extends EventEmitter implements FileStorage {
   bucket: string;
-  private s3: awsSdk.S3;
+  s3: S3;
 
   constructor(options: S3StorageOptions) {
     super();
-    this.s3 = new awsSdk.S3(Object.assign({ apiVersion: '2006-03-01' }, options.clientOptions));
+    const opts = {
+      apiVersion: '2006-03-01',
+      ...(options.endpoint ? { endpoint: options.endpoint, s3BucketEndpoint: true } : {})
+    };
+    this.s3 = new AWS.S3(Object.assign(opts, options.clientOptions));
     this.bucket = options.bucket;
+    if (!enabled) {
+      return;
+    }
+    // eslint-disable-next-line @typescript-eslint/no-empty-function
     this.initialize().then(() => {});
   }
 
@@ -34,13 +45,13 @@ export class S3Storage extends EventEmitter implements FileStorage {
     }
   }
 
-  async list(id: string): Promise<(Directory | File)[]> {
+  async list(_id: string): Promise<(Directory | File)[]> {
     const data = await this.s3.listObjects({ Bucket: this.bucket }).promise();
-    return data.Contents!.map((obj: s3.Object) => {
+    return data.Contents!.map((obj: S3.Object) => {
       return {
         type: obj.Key!.endsWith('/') ? ItemType.Directory : ItemType.File,
-        name: obj.Key || '',
-        refId: obj.Key || '',
+        name: obj.Key || '',
+        refId: obj.Key || '',
       };
     });
   }
@@ -57,7 +68,7 @@ export class S3Storage extends EventEmitter implements FileStorage {
 
   async getContent(file: File): Promise<Buffer> {
     const data = await this.s3.getObject({ Bucket: this.bucket, Key: file.refId }).promise();
-    return _.isString(data.Body!) ? new Buffer(data.Body! as string) : data.Body! as Buffer;
+    return _.isString(data.Body!) ? Buffer.from(data.Body! as string) : data.Body! as Buffer;
   }
 
   async removeFile(file: File): Promise<void> {
@@ -83,7 +94,7 @@ export class S3Storage extends EventEmitter implements FileStorage {
     return { name: item.name, type: ItemType.File, refId: key, mimetype: item.mimetype };
   }
 
-  async moveDirectory(dir: Directory, destination: Directory, refCb: (oldItem: Item, newItem: Item) => Promise<void>): Promise<Directory> {
+  async moveDirectory(dir: Directory, destination: Directory): Promise<MoveDirectoryResponse> {
     const targetRef = destination.refId ? destination.refId : `${destination.name}/`;
     const data = await this.s3.listObjects({ Bucket: this.bucket, Prefix: dir.refId }).promise();
     const exists = await this.s3.listObjects({ Bucket: this.bucket, Prefix: targetRef }).promise();
@@ -94,7 +105,10 @@ export class S3Storage extends EventEmitter implements FileStorage {
       },
     };
     const targetPrefix = exists.Contents!.length ? `${targetRef}${dir.name}/` : `${targetRef}`;
-    await Promise.all(_.map(data.Contents!, async (s3item: s3.Object) => {
+    const items: {
+      oldItem: Item;
+      newItem: Item;
+    }[] = await Promise.all(_.map(data.Contents!, async (s3item: S3.Object) => {
       deleteParams.Delete.Objects.push({ Key: s3item.Key! });
       const newKey = s3item.Key!.replace(dir.refId, targetPrefix);
       await this.s3.copyObject({
@@ -102,19 +116,22 @@ export class S3Storage extends EventEmitter implements FileStorage {
         CopySource: `${this.bucket}/${s3item.Key}`,
         Key: newKey,
       }).promise();
-      if (s3item.Key === dir.refId) return;
-      return refCb({
-        refId: s3item.Key!,
-        name: path.basename(s3item.Key!),
-        type: s3item.Key!.endsWith('/') ? ItemType.Directory : ItemType.File,
-      },           {
+      return {
+        oldItem: {
+          refId: s3item.Key!,
+          name: path.basename(s3item.Key!),
+          type: s3item.Key!.endsWith('/') ? ItemType.Directory : ItemType.File,
+        },
+        newItem: {
         refId: newKey,
         name: path.basename(s3item.Key!),
         type: s3item.Key!.endsWith('/') ? ItemType.Directory : ItemType.File,
-      });
+      }};
     }));
     await this.s3.deleteObjects(deleteParams).promise();
-    return { name: path.basename(targetPrefix), type: ItemType.Directory, refId: targetPrefix };
+    return {
+      directory: { name: path.basename(targetPrefix), type: ItemType.Directory, refId: targetPrefix },
+      items
+    };
   }
-
 }
