@@ -1,4 +1,4 @@
-import * as _ from 'lodash';
+import { split } from 'lodash';
 
 export interface FilesystemStorageOptions {
   rootDir: string;
@@ -13,32 +13,33 @@ export interface S3StorageOptions {
     region: string;
   };
   bucket: string;
+  endpoint: string;
 }
 
 export const config = {
   storage: {
     enabled: process.env.STORAGE_ENABLED || 'filesystem',
     filesystem: {
-      rootDir: process.env.FILESYSTEM_STORAGE_ROOT || '/tmp/filebank',
+      rootDir: process.env.FILESYSTEM_STORAGE_ROOT || '/tmp/filebank',
     } as FilesystemStorageOptions,
     s3: {
       clientOptions: {
         credentials: {
-          accessKeyId: process.env.S3_STORAGE_ACCESS_KEY || '',
+          accessKeyId: process.env.S3_STORAGE_ACCESS_KEY || '',
           secretAccessKey: process.env.S3_STORAGE_SECRET_KEY || '',
         },
-        region: process.env.S3_STORAGE_REGION || 'eu-west-1',
+        region: process.env.S3_STORAGE_REGION || 'eu-west-1',
       },
-      bucket: process.env.S3_STORAGE_BUCKET || 'filebank',
+      bucket: process.env.S3_STORAGE_BUCKET || 'filebank',
     } as S3StorageOptions,
   },
   schemaRequired: !!process.env.SCHEMA_REQUIRED,
   jwtKey: process.env.JWT_KEY || 'v3rys3cr3tK3y',
   authz: {
-    enabled: JSON.parse(process.env.JWT_ENABLED || 'true'),
-    readScope: _.split(process.env.JWT_READ_SCOPE || 'filebank:read', ','),
-    writeScope: _.split(process.env.JWT_WRITE_SCOPE || 'filebank:write', ','),
-    deleteScope: _.split(process.env.JWT_DELETE_SCOPE || 'filebank:delete', ','),
+    enabled: JSON.parse(process.env.JWT_ENABLED || 'true'),
+    readScope: split(process.env.JWT_READ_SCOPE || 'filebank:read', ','),
+    writeScope: split(process.env.JWT_WRITE_SCOPE || 'filebank:write', ','),
+    deleteScope: split(process.env.JWT_DELETE_SCOPE || 'filebank:delete', ','),
   },
-  mongoDbUrl: process.env.MONGODB_URL || 'mongodb://localhost:27017/filebank',
+  mongoDbUrl: process.env.MONGODB_URL || 'mongodb://localhost:27017/filebank',
 };
