@@ -2,7 +2,7 @@ import { directory } from './directory';
 import { file } from './file';
 // import * as _ from 'lodash';
 
-export default {
-  Directory: directory,
-  File: file,
-};
+// tslint:disable-next-line:variable-name
+export const Directory = directory;
+// tslint:disable-next-line:variable-name
+export const File = file;
