@@ -1,9 +1,9 @@
 import fs = require('fs');
 import path = require('path');
 import fsExtra = require('fs-extra');
-import * as glob from 'glob';
-import { Directory, File, FileStorage, Item, ItemType } from '../index';
-import * as Bluebird from 'bluebird';
+import glob from 'glob';
+import { promisify } from "util";
+import { Directory, File, FileStorage, Item, ItemType, MoveDirectoryResponse } from "../index";
 
 const DEFAULT_MIMETYPE = 'application/octet-stream';
 
@@ -37,8 +37,8 @@ export class FilesystemStorage implements FileStorage {
   }
   async list(id: string): Promise<(Directory | File)[]> {
     const dirPath : string = path.resolve(this.rootDir, id);
-    const filenames: string[] = await Bluebird.fromNode<string[]>(cb => fs.readdir(dirPath, cb));
-    const items: (Directory | File)[] = [];
+    const filenames: string[] = fs.readdirSync(dirPath);
+    const items: (Directory | File)[] = [];
     filenames.map((filename) => {
       const ref = this.relatify(path.resolve(this.rootDir, id, filename));
       if (fs.statSync(path.resolve(dirPath, filename)).isDirectory()) {
@@ -60,8 +60,8 @@ export class FilesystemStorage implements FileStorage {
   }
 
   async mkdir(dirRef: string, dir: Directory): Promise<Directory> {
-    const dirPath = path.resolve(this.rootDir, dirRef || '', dir.name);
-    await Bluebird.fromNode(cb => fs.mkdir(dirPath, cb));
+    const dirPath = path.resolve(this.rootDir, dirRef || "", dir.name);
+    fs.mkdirSync(dirPath);
     return {
       refId: this.relatify(dirPath),
       name: dir.name,
@@ -79,16 +79,16 @@ export class FilesystemStorage implements FileStorage {
   }
 
   async getContent(file: File): Promise<Buffer> {
-    return Bluebird.fromNode<Buffer>(cb => fs.readFile(path.resolve(this.rootDir, file.refId), cb));
+    return fs.readFileSync(path.resolve(this.rootDir, file.refId));
   }
 
   async removeFile(file: File): Promise<void> {
-    await Bluebird.fromNode(cb => fs.unlink(path.resolve(this.rootDir, file.refId), cb));
+    fs.unlinkSync(path.resolve(this.rootDir, file.refId));
   }
 
   async createFile(file: File, directory: Directory|undefined, data: Buffer): Promise<File> {
     const filePath = path.resolve(this.rootDir, directory ? directory.refId : '', file.name);
-    await Bluebird.fromCallback(cb => fs.writeFile(filePath, data, cb));
+    fs.writeFileSync(filePath, data);
     return {
       mimetype: DEFAULT_MIMETYPE,
       refId: this.relatify(filePath),
@@ -98,19 +98,8 @@ export class FilesystemStorage implements FileStorage {
   }
 
   async moveFile(file: File, destination: Directory): Promise<File> {
-    function move(oldPath: string, newPath: string, callback: (err?: NodeJS.ErrnoException) => void) {
-      fs.rename(oldPath, newPath, (err) => {
-        if (err) {
-          if (err.code === 'EXDEV') {
-            copy();
-          } else {
-            callback(err);
-          }
-          return;
-        }
-        callback();
-      });
-      function copy() {
+    function move(oldPath: string, newPath: string, callback: (err?: any) => void) {
+      const copy = () => {
         const readStream = fs.createReadStream(oldPath);
         const writeStream = fs.createWriteStream(newPath);
 
@@ -122,27 +111,44 @@ export class FilesystemStorage implements FileStorage {
         });
 
         readStream.pipe(writeStream);
-      }
+      };
+      fs.rename(oldPath, newPath, (err) => {
+        if (err) {
+          if (err.code === 'EXDEV') {
+            copy();
+          } else {
+            callback(err);
+          }
+          return;
+        }
+        callback();
+      });
     }
+
+    const moveAsync = promisify(move);
 
     const oldPath = path.resolve(this.rootDir, file.refId);
     const filename = file.refId.split('/').pop();
     const destPath = path.resolve(this.rootDir, destination.refId, filename || '');
-    await Bluebird.fromCallback(cb => move(oldPath, destPath, cb));
+    await moveAsync(oldPath, destPath);
     file.refId = this.relatify(destPath);
     return file;
   }
 
-  async moveDirectory(dir: Directory, destination: Directory, refCb: (oldItem: Item, newItem: Item) => Promise<void>): Promise<Directory> {
+  async moveDirectory(dir: Directory, destination: Directory): Promise<MoveDirectoryResponse> {
     const subdirRoot = path.resolve(this.rootDir, dir.refId);
-    const items = await Bluebird.fromNode<string[]>(cb => glob('**', { cwd: subdirRoot }, cb));
+    const globAsync = promisify(glob);
+    const items = await globAsync('**', { cwd: subdirRoot });
     const destinationRef = destination.refId ? destination.refId : destination.name;
     const targetPath = fs.existsSync(path.resolve(this.rootDir, destinationRef)) ?
       path.resolve(this.rootDir, destinationRef, path.basename(dir.refId)) :
       path.resolve(this.rootDir, destinationRef);
 
     fsExtra.moveSync(subdirRoot, targetPath);
-    items.map((item: string) => {
+    const allItems: {
+      oldItem: Item;
+      newItem: Item;
+    }[] = items.map((item: string) => {
       const stat = fs.statSync(path.resolve(targetPath, item));
       const oldItem = { name: path.basename(item), refId: path.join(dir.refId, item), type: stat.isDirectory() ? ItemType.Directory : ItemType.File };
       const newItem = {
@@ -150,9 +156,15 @@ export class FilesystemStorage implements FileStorage {
         refId: path.join(destinationRef, item),
         type: stat.isDirectory() ? ItemType.Directory : ItemType.File,
       };
-      return refCb(oldItem, newItem);
+      return {
+        oldItem,
+        newItem
+      };
     });
-    return { type: ItemType.Directory, name: path.basename(targetPath), refId: this.relatify(targetPath) };
+    return {
+      directory: { type: ItemType.Directory, name: path.basename(targetPath), refId: this.relatify(targetPath) },
+      items: allItems
+    };
   }
 
   setRootDir(dir: string) {
